use manage.py to run tests
